Feature/task3

+++ Models/BaseModel.cs
@@ -10,7 +10,7 @@ namespace paroquiaRussas.Models
 {
     public abstract class BaseModel
     {
-        public int Id { get; set; }
+        public long Id { get; set; }
 
         public static void Configure(ModelBuilder modelBuilder)
         {
@@ -19,7 +19,7 @@ namespace paroquiaRussas.Models
                 if (typeof(BaseModel).IsAssignableFrom(entityType.ClrType))
                 {
                     var entity = modelBuilder.Entity(entityType.ClrType);
-                    entity.Property<int>("Id").HasColumnName($"{entityType.ClrType.Name}Id");
+                    entity.Property<long>("Id").HasColumnName($"{entityType.ClrType.Name}Id");
                 }
             }
         }
